api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via starlette
beautifulsoup4==4.14.3
    # via yfinance
certifi==2026.1.4
    # via
    #   curl-cffi
    #   requests
cffi==2.0.0
    # via
    #   cryptography
    #   curl-cffi
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
cryptography==46.0.5
    # via google-auth
curl-cffi==0.13.0
    # via yfinance
fastapi==0.129.0
    # via -r requirements.in
frozendict==2.4.7
    # via yfinance
google-api-core[grpc]==2.29.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
google-auth==2.48.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-core
google-cloud-bigquery==3.40.1
    # via -r requirements.in
google-cloud-core==2.5.0
    # via google-cloud-bigquery
google-crc32c==1.8.0
    # via google-resumable-media
google-resumable-media==2.8.0
    # via google-cloud-bigquery
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.78.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.78.0
    # via google-api-core
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
multitasking==0.0.12
    # via yfinance
numpy==2.4.2
    # via
    #   -r requirements.in
    #   pandas
    #   scipy
    #   yfinance
packaging==26.0
    # via google-cloud-bigquery
pandas==3.0.1
    # via yfinance
peewee==3.19.0
    # via yfinance
platformdirs==4.9.2
    # via yfinance
proto-plus==1.27.1
    # via google-api-core
protobuf==6.33.5
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   yfinance
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via
    #   google-cloud-bigquery
    #   pandas
pytz==2025.2
    # via yfinance
requests==2.32.5
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   yfinance
rsa==4.9.1
    # via google-auth
scipy==1.17.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
soupsieve==2.8.3
    # via beautifulsoup4
starlette==0.52.1
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
urllib3==2.6.3
    # via requests
uvicorn==0.41.0
    # via -r requirements.in
websockets==16.0
    # via yfinance
yfinance==1.2.0
    # via -r requirements.in
